backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.32
alembic==1.13.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.8.2
python-jose==3.5.0
cryptography==45.0.7
requests==2.32.5